+++ CHANGELOG.md
@@ -36,6 +36,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Deprecated `colors.search.bar`, use `colors.footer_bar` instead
 - On macOS, Alacritty now reads `AppleFontSmoothing` from user defaults to control font smoothing
 - Show warning when either `columns` or `lines` is set, but not both
+- Warn when either `columns` or `lines` is non-zero, but not both
 
 ### Fixed